Fix bug for mixed simulator (#63)
* _expand state complex 128 to complex 64

* Update Changelog.

* Update Changelog.

+++ pennylane_cirq/simulator_device.py
@@ -118,7 +118,7 @@ class SimulatorDevice(CirqDevice):
         # get indices for which the state is changed to input state vector elements
         ravelled_indices = np.ravel_multi_index(unravelled_indices.T, [2] * self.num_wires)
 
-        state = np.zeros([2 ** self.num_wires], dtype=np.complex128)
+        state = np.zeros([2 ** self.num_wires], dtype=np.complex64)
         state[ravelled_indices] = state_vector
         state_vector = state.reshape([2] * self.num_wires)
